+++ src/rigid.rs
@@ -5,39 +5,39 @@ use crate::frame::Frame;
 #[derive(Debug, Clone, Copy)]
 pub struct Rigid {
     /// Mass in `kg`
-    pub mass: f32,
+    pub mass: f64,
 
     /// Rotational inertia in `kg m^2`
-    pub rotational_inertia: Vector3<f32>,
+    pub rotational_inertia: Vector3<f64>,
 
     /// Force acting on the rigid body outside its frame.
     /// Measured in `N`.
-    pub external_force: Vector3<f32>,
+    pub external_force: Vector3<f64>,
 
     /// Force acting on the rigid body within its own frame.
     /// Measured in `N`.
-    pub internal_force: Vector3<f32>,
+    pub internal_force: Vector3<f64>,
 
     /// Torque acting on the rigid body outside its frame.
     /// Measrued in `N m`.
-    pub external_torque: Vector3<f32>,
+    pub external_torque: Vector3<f64>,
 
     /// Torque acting on the rigid body within its own frame.
     /// Measrued in `N m`.
-    pub internal_torque: Vector3<f32>,
+    pub internal_torque: Vector3<f64>,
 
     /// Current velocity of the rigid body in `m s^-1`
-    pub velocity: Vector3<f32>,
+    pub velocity: Vector3<f64>,
 
     /// Current angular velocity of the rigid body in `s^-1`
-    pub angular_velocity: Vector3<f32>,
+    pub angular_velocity: Vector3<f64>,
 
     pub frame: Frame,
     pub past_frame: Frame,
 }
 
 impl Rigid {
-    pub fn new(mass: f32) -> Rigid {
+    pub fn new(mass: f64) -> Rigid {
         let extent = Vector3::new(1.0, 1.0, 1.0);
         let inertia = 1.0 / 12.0
             * mass
@@ -61,7 +61,7 @@ impl Rigid {
         }
     }
 
-    pub fn integrate(&mut self, dt: f32) {
+    pub fn integrate(&mut self, dt: f64) {
         let force = self.external_force + self.frame.quaternion * self.internal_force;
         self.velocity += dt * force / self.mass;
 
@@ -74,13 +74,13 @@ impl Rigid {
             .integrate(dt, self.velocity, self.angular_velocity);
     }
 
-    pub fn derive(&mut self, dt: f32) {
+    pub fn derive(&mut self, dt: f64) {
         (self.velocity, self.angular_velocity) = self.frame.derive(dt, self.past_frame)
     }
 
     /// Applies a linear impulse in a given direction and magnitude at a given location.
     /// Results in changes in both position and quaternion.
-    pub fn apply_impulse(&mut self, impulse: Vector3<f32>, point: Vector3<f32>) {
+    pub fn apply_impulse(&mut self, impulse: Vector3<f64>, point: Vector3<f64>) {
         self.frame.position += impulse / self.mass;
 
         let log = (point - self.frame.position)
@@ -91,7 +91,7 @@ impl Rigid {
     }
 
     /// Computes the position difference of a global point in the current frame from the same point in the past frame.
-    pub fn delta(&self, global: Vector3<f32>) -> Vector3<f32> {
+    pub fn delta(&self, global: Vector3<f64>) -> Vector3<f64> {
         let local = self.frame.inverse().act(global);
         let past_global = self.past_frame.act(local);
         global - past_global
